+++ packages/eslint-config-ts/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @brucesong/eslint-config-ts
 
+## 0.0.6
+
+### Patch Changes
+
+- Update.
+- Updated dependencies
+  - @brucesong/eslint-config-basic@0.0.6
+
 ## 0.0.5
 
 ### Patch Changes